requirements.txt
flask==1.1.2
torch==1.6.0
torchvision==0.6.1
pillow==7.2.0
